+++ packages/viz/webpack.config.js
@@ -126,6 +126,8 @@ const resolve = {
     "cartridge-vicentra.png": path.resolve(__dirname, "../../branding/sitechange/cartridge-vicentra.png"),
     "infusion.png": path.resolve(__dirname, "../../branding/sitechange/infusion.png"),
     "warmup-dexcom.svg": path.resolve(__dirname, "../../branding/warmup/warmup-dexcom.svg"),
+    "jaFont-Regular.ttf": path.resolve(__dirname, "../../branding/fonts/noto/NotoSerifJP-Regular.ttf"),
+    "jaFont-Bold.ttf": path.resolve(__dirname, "../../branding/fonts/noto/NotoSerifJP-Bold.ttf"),
   },
   fallback: {
     stream: require.resolve("stream-browserify"),