+++ packages/yourloops/test/pages/notifications/notifications-list.test.tsx
@@ -26,7 +26,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { NotificationsPage } from "../../../pages/notifications/notifications-list";
+import NotificationsPage from "../../../pages/notifications/notifications-list";
 
 describe("Notification list", () => {
   it("should be exported as a function", () => {